+++ packages/orm/src/codegen/appCodeUpdater.ts
@@ -4,6 +4,7 @@ import { updateMainFile } from './updateMainFile';
 import { updateTableFile } from './updateTableFile/updateTableFile';
 import { createBaseTableFile } from './createBaseTableFile';
 import { SetOptional } from 'orchid-core';
+import { QueryLogOptions } from 'pqb';
 
 export class AppCodeUpdaterError extends Error {}
 
@@ -12,6 +13,7 @@ export type AppCodeUpdaterConfig = {
   baseTablePath: string;
   baseTableName: string;
   mainFilePath: string;
+  logger?: QueryLogOptions['logger'];
 };
 
 export const appCodeUpdater = ({
@@ -20,7 +22,7 @@ export const appCodeUpdater = ({
   baseTableName,
   mainFilePath,
 }: SetOptional<AppCodeUpdaterConfig, 'baseTableName'>): AppCodeUpdater => {
-  return async ({ ast, options, basePath, cache: cacheObject }) => {
+  return async ({ ast, options, basePath, cache: cacheObject, logger }) => {
     const params: AppCodeUpdaterConfig = {
       tablePath(name: string) {
         const file = tablePath(name);
@@ -29,10 +31,17 @@ export const appCodeUpdater = ({
       baseTablePath: resolvePath(basePath, baseTablePath),
       baseTableName: baseTableName || 'BaseTable',
       mainFilePath: resolvePath(basePath, mainFilePath),
+      logger,
     };
 
     const promises: Promise<void>[] = [
-      updateMainFile(params.mainFilePath, params.tablePath, ast, options),
+      updateMainFile(
+        params.mainFilePath,
+        params.tablePath,
+        ast,
+        options,
+        logger,
+      ),
       updateTableFile({ ...params, ast }),
     ];