+++ packages/orm/src/codegen/updateTableFile/createTable.ts
@@ -1,7 +1,7 @@
 import { RakeDbAst } from 'rake-db';
 import { getImportPath } from '../utils';
 import { codeToString, columnsShapeToCode } from 'pqb';
-import { Code, singleQuote } from 'orchid-core';
+import { Code, singleQuote, pathToLog } from 'orchid-core';
 import { toPascalCase } from '../../utils';
 import fs from 'fs/promises';
 import { UpdateTableFileParams } from './updateTableFile';
@@ -9,6 +9,7 @@ import path from 'path';
 
 export const createTable = async ({
   ast,
+  logger,
   ...params
 }: UpdateTableFileParams & { ast: RakeDbAst.Table }) => {
   const tablePath = params.tablePath(ast.name);
@@ -44,6 +45,7 @@ export const createTable = async ({
   await fs.mkdir(path.dirname(tablePath), { recursive: true });
   try {
     await fs.writeFile(tablePath, codeToString(code, '', '  '), { flag: 'wx' });
+    logger?.log(`Created ${pathToLog(tablePath)}`);
   } catch (err) {
     if ((err as unknown as { code: string }).code !== 'EEXIST') {
       throw err;