+++ packages/orm/src/codegen/updateTableFile/renameTable.test.ts
@@ -1,7 +1,8 @@
 import { updateTableFile } from './updateTableFile';
 import { asMock, ast, makeTestWritten, tablePath } from '../testUtils';
-import path from 'path';
+import { resolve } from 'path';
 import fs from 'fs/promises';
+import { pathToLog } from 'orchid-core';
 
 jest.mock('fs/promises', () => ({
   mkdir: jest.fn(),
@@ -9,11 +10,22 @@ jest.mock('fs/promises', () => ({
   writeFile: jest.fn(),
 }));
 
-const baseTablePath = path.resolve('baseTable.ts');
+const baseTablePath = resolve('baseTable.ts');
 const baseTableName = 'BaseTable';
-const params = { baseTablePath, baseTableName, tablePath };
+const log = jest.fn();
+const params = {
+  baseTablePath,
+  baseTableName,
+  tablePath,
+  logger: { ...console, log },
+};
 
-const testWritten = makeTestWritten(tablePath('some'));
+const path = tablePath('some');
+const testWrittenOnly = makeTestWritten(path);
+const testWritten = (content: string) => {
+  testWrittenOnly(content);
+  expect(log).toBeCalledWith(`Updated ${pathToLog(path)}`);
+};
 
 describe('renameTable', () => {
   beforeEach(() => {